requirements.txt
boto3==1.28.42
botocore==1.31.42
greenlet==2.0.2
jmespath==1.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.6.2
six==1.16.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
urllib3==1.26.16
watchdog==3.0.0
